PRD-Maker-with-LLM-feature-llm-service2/PRD-Maker-with-LLM-feature-llm-service/prd_generator/requirements.txt
google-generativeai==0.3.1
python-dotenv==1.0.0
openai==1.3.7